os: rename aio.h to avoid conflict
FreeBSD's libc offers <aio.h>.
